convert red flags to a blueprint
