+++ app/red_flags/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+red_flag_bp = Blueprint('red_flag', __name__, url_prefix='/api/v1')
